Added Mailer.send_info_to_interrogator
Right after the game starts, the system sends to each player all the
information about the other players. This methods sends the names, email
and descriptions of the impostor and the honest player to the
interrogator, the user that started the game.

Now there is one Mailer object per game, to handle all the email for it.
This implies that the constructor now optionally takes a game object.