Give aliases to players.
The alias is the mask with which the game presents the players to the
interrogator - in every game, either the impostor or the honest player
will be player A and the other one will be player B. Essentially, the
game consists of discovering the player behind the alias.

This change implied linking answers to players and refactoring the code
to follow the new db structure.